requirements.txt
torch>=2.2
torch-geometric>=2.5
numpy>=1.26
pandas>=2.0
pyyaml>=6.0
opencv-python>=4.9
matplotlib>=3.8